Fix bug 6513826: Add less verbose logging level for MBean messages. Raised the level of the javax.management package to SEVERE, matching the other core packages (such as javax.net and org.springframework).
The gsac project used an older copy of the file system connector's
logging.properties, carelessly merged with the SharePoint connector's
version of the file (e.g., with duplicate lines, and loggers
configured in two different places in the file). I integrated them and
updated some whitespace to match the latest versions from the two
connectors.

This change is only applied to the "offboard" copy of
logging.properties. That's the one that is actually used, and the
"onboard" copy is broken in other ways. See bug 7241822 for details.

Code review: https://codereview.appspot.com/6570057